+++ support-frontend/assets/pages/weekly-subscription-checkout/weeklySubscriptionCheckout.tsx
@@ -1,7 +1,6 @@
 // ----- Imports ----- //
 import { FocusStyleManager } from '@guardian/source-foundations';
 import { Provider } from 'react-redux';
-import WeeklyFooter from 'components/footerCompliant/WeeklyFooter';
 import Page from 'components/page/page';
 import HeaderWrapper from 'components/subscriptionCheckouts/headerWrapper';
 import CheckoutStage from 'components/subscriptionCheckouts/stage';
@@ -10,17 +9,15 @@ import type { WeeklyBillingPeriod } from 'helpers/productPrice/billingPeriods';
 import { postIntroductorySixForSixBillingPeriod } from 'helpers/productPrice/billingPeriods';
 import { Domestic } from 'helpers/productPrice/fulfilmentOptions';
 import { NoProductOptions } from 'helpers/productPrice/productOptions';
-import { getProductPrice } from 'helpers/productPrice/productPrices';
-import { getAppliedPromo } from 'helpers/productPrice/promotions';
 import { GuardianWeekly } from 'helpers/productPrice/subscriptions';
 import type { CommonState } from 'helpers/redux/commonState/state';
 import { initReduxForSubscriptions } from 'helpers/redux/subscriptionsStore';
 import { renderPage } from 'helpers/rendering/render';
 import { createWithDeliveryCheckoutReducer } from 'helpers/subscriptionsForms/subscriptionCheckoutReducer';
-import { promotionTermsUrl } from 'helpers/urls/routes';
 import { getQueryParameter } from 'helpers/urls/url';
 import { formatMachineDate } from 'helpers/utilities/dateConversions';
 import { getWeeklyDays } from 'pages/weekly-subscription-checkout/helpers/deliveryDays';
+import { GuardianWeeklyFooter } from '../../components/footerCompliant/FooterWithPromoTerms';
 import ThankYouContent from './components/thankYou';
 import WeeklyCheckoutForm from './components/weeklyCheckoutForm';
 import WeeklyCheckoutFormGifting from './components/weeklyCheckoutFormGifting';
@@ -49,33 +46,21 @@ const reducer = (commonState: CommonState) =>
 	);
 
 const store = initReduxForSubscriptions(reducer);
-const {
-	orderIsAGift,
-	billingPeriod,
-	productPrices,
-	fulfilmentOption,
-	productOption,
-} = store.getState().page.checkout;
+const { orderIsAGift, productPrices } = store.getState().page.checkout;
 const { countryId } = store.getState().common.internationalisation;
-const productPrice = getProductPrice(
-	productPrices,
-	countryId,
-	billingPeriod,
-	fulfilmentOption,
-	productOption,
-);
-const appliedPromo = getAppliedPromo(productPrice.promotions);
-const defaultPromo = orderIsAGift ? 'GW20GIFT1Y' : '10ANNUAL';
-const promoTermsLink = promotionTermsUrl(
-	appliedPromo ? appliedPromo.promoCode : defaultPromo,
-);
 FocusStyleManager.onlyShowFocusOnTabs();
 // ----- Render ----- //
 const content = (
 	<Provider store={store}>
 		<Page
 			header={<HeaderWrapper />}
-			footer={<WeeklyFooter promoTermsLink={promoTermsLink} />}
+			footer={
+				<GuardianWeeklyFooter
+					productPrices={productPrices}
+					orderIsAGift={!!orderIsAGift}
+					country={countryId}
+				/>
+			}
 		>
 			<CheckoutStage
 				checkoutForm={
