+++ support-frontend/assets/pages/weekly-subscription-landing/weeklySubscriptionLanding.tsx
@@ -1,7 +1,6 @@
 // ----- Imports ----- //
 import CentredContainer from 'components/containers/centredContainer';
 import FullWidthContainer from 'components/containers/fullWidthContainer';
-import WeeklyFooter from 'components/footerCompliant/WeeklyFooter';
 import headerWithCountrySwitcherContainer from 'components/headers/header/headerWithCountrySwitcher';
 import Block from 'components/page/block';
 import Page from 'components/page/page';
@@ -18,14 +17,11 @@ import {
 	UnitedStates,
 } from 'helpers/internationalisation/countryGroup';
 import { setUpTrackingAndConsents } from 'helpers/page/page';
-import {
-	getPromotionCopy,
-	promoQueryParam,
-} from 'helpers/productPrice/promotions';
+import { getPromotionCopy } from 'helpers/productPrice/promotions';
 import { renderPage } from 'helpers/rendering/render';
-import { promotionTermsUrl, routes } from 'helpers/urls/routes';
+import { routes } from 'helpers/urls/routes';
 import 'stylesheets/skeleton/skeleton.scss';
-import { getQueryParameter } from 'helpers/urls/url';
+import { GuardianWeeklyFooter } from '../../components/footerCompliant/FooterWithPromoTerms';
 import Benefits from './components/content/benefits';
 import GiftBenefits from './components/content/giftBenefits';
 import { WeeklyHero } from './components/hero/hero';
@@ -54,6 +50,10 @@ function WeeklyLandingPage({
 	countryGroupId,
 	participations,
 }: WeeklyLandingPropTypes) {
+	if (!productPrices) {
+		return null;
+	}
+
 	const path = orderIsAGift
 		? routes.guardianWeeklySubscriptionLandingGift
 		: routes.guardianWeeklySubscriptionLanding;
@@ -61,13 +61,6 @@ function WeeklyLandingPage({
 		? routes.guardianWeeklySubscriptionLanding
 		: routes.guardianWeeklySubscriptionLandingGift;
 	const sanitisedPromoCopy = getPromotionCopy(promotionCopy);
-	const defaultPromo = (): string => {
-		if (orderIsAGift) return 'GW20GIFT1Y';
-		return '10ANNUAL';
-	};
-	const promoTermsLink = promotionTermsUrl(
-		getQueryParameter(promoQueryParam, defaultPromo()),
-	);
 	// ID for Selenium tests
 	const pageQaId = `qa-guardian-weekly${orderIsAGift ? '-gift' : ''}`;
 	const Header = headerWithCountrySwitcherContainer({
@@ -88,7 +81,13 @@ function WeeklyLandingPage({
 		<Page
 			id={pageQaId}
 			header={<Header />}
-			footer={<WeeklyFooter centred promoTermsLink={promoTermsLink} />}
+			footer={
+				<GuardianWeeklyFooter
+					productPrices={productPrices}
+					orderIsAGift={!!orderIsAGift}
+					country={countryId}
+				/>
+			}
 		>
 			<WeeklyHero
 				orderIsAGift={orderIsAGift ?? false}
